requirements.txt
flask==2.3.3
yfinance==0.2.28
pandas==2.0.3
numpy==1.24.3
psycopg2-binary==2.9.7
requests==2.31.0
python-dotenv==1.0.0
tabulate==0.9.0
